FIXME: Add core `OwnRef<'_, dyn Trait>` functionality
TODO: after thinking about it, it is too ambitious, and thus, too limiting,
to want `OwnRef<impl Trait> : Trait`.

We should rather produce a companion extension trait, say `TraitOwnRef`, with
only the `dyn`-safe `self`-by-value methods in it, and move all of our already existing
logic to handle that reduced safe subset of use cases.

  - To avoid method clashes, these methods would have to be renamed, e.g. with an `_ownref` suffix.

Then, the user of the macro is always free to write their own blanket impl using the `…_ownref()` helpers.

  - If they get blanket impl errors, they will know why;
  - No more problems with default impls / fn bodies;
  - No more issues with assoc types;
  - etc.

The only drawback is the "uglyness" of both involving a `…_ownref()`-renamed method, and of an `…OwnRef` extra
trait (the extension trait).